generate private destructor declaration for non-instantiable classes
Originally, gcc would happily generate a (broken) destructor in these
cases, but under modern C++ standards, it deletes the destructor and
throws an error. By delaring the destructor, but not implementing it, we
apparently block that gcc code path. The lack of an implementation
should not be a problem, since we should never instantiate these wrapper
classes.